+++ src/main/java/com/gandec/ganadecs/Repository/PropietariosRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PropietariosRepository extends JpaRepository<Propietario,Integer> {
+public interface PropietariosRepository extends JpaRepository<Propietario,Long> {
 }
